Extract relevant okhttp client code to AbstractOkHttpClient
